[DUBBO-2988] Unrecognized the other provider (#3013)
*     #2988 [Dubbo - telnet] Unrecognized the other provider

* remove useless files added by accident.

* fix cycle dependency

* fix more cycle dependency

* revert unnecessary change in demo

* make unit test pass

* revert LF to the original CRLF

* change to use lambda

* fix cycle dependency

* remove useless imports

* make unit test stable

* make refactor change according to code review comments

* refactor as review comments suggest.
