+++ dubbo-rpc/dubbo-rpc-dubbo/src/main/java/org/apache/dubbo/rpc/protocol/dubbo/telnet/ListTelnetHandler.java
@@ -18,14 +18,20 @@ package org.apache.dubbo.rpc.protocol.dubbo.telnet;
 
 import org.apache.dubbo.common.extension.Activate;
 import org.apache.dubbo.common.utils.ReflectUtils;
+import org.apache.dubbo.common.utils.StringUtils;
 import org.apache.dubbo.remoting.Channel;
 import org.apache.dubbo.remoting.telnet.TelnetHandler;
 import org.apache.dubbo.remoting.telnet.support.Help;
-import org.apache.dubbo.rpc.Exporter;
-import org.apache.dubbo.rpc.Invoker;
-import org.apache.dubbo.rpc.protocol.dubbo.DubboProtocol;
+import org.apache.dubbo.rpc.model.ApplicationModel;
+import org.apache.dubbo.rpc.model.ConsumerMethodModel;
+import org.apache.dubbo.rpc.model.ConsumerModel;
+import org.apache.dubbo.rpc.model.ProviderMethodModel;
+import org.apache.dubbo.rpc.model.ProviderModel;
 
 import java.lang.reflect.Method;
+
+import static org.apache.dubbo.registry.support.ProviderConsumerRegTable.getConsumerAddressNum;
+import static org.apache.dubbo.registry.support.ProviderConsumerRegTable.isRegistered;
 
 /**
  * ListTelnetHandler
@@ -45,7 +51,7 @@ public class ListTelnetHandler implements TelnetHandler {
                 if ("-l".equals(part)) {
                     detail = true;
                 } else {
-                    if (service != null && service.length() > 0) {
+                    if (!StringUtils.isEmpty(service)) {
                         return "Invalid parameter " + part;
                     }
                     service = part;
@@ -54,47 +60,95 @@ public class ListTelnetHandler implements TelnetHandler {
         } else {
             service = (String) channel.getAttribute(ChangeTelnetHandler.SERVICE_KEY);
             if (service != null && service.length() > 0) {
-                buf.append("Use default service " + service + ".\r\n");
+                buf.append("Use default service ").append(service).append(".\r\n");
             }
         }
-        if (service == null || service.length() == 0) {
-            for (Exporter<?> exporter : DubboProtocol.getDubboProtocol().getExporters()) {
-                if (buf.length() > 0) {
-                    buf.append("\r\n");
-                }
-                buf.append(exporter.getInvoker().getInterface().getName());
-                if (detail) {
-                    buf.append(" -> ");
-                    buf.append(exporter.getInvoker().getUrl());
-                }
-            }
+
+        if (StringUtils.isEmpty(service)) {
+            printAllServices(buf, detail);
         } else {
-            Invoker<?> invoker = null;
-            for (Exporter<?> exporter : DubboProtocol.getDubboProtocol().getExporters()) {
-                if (service.equals(exporter.getInvoker().getInterface().getSimpleName())
-                        || service.equals(exporter.getInvoker().getInterface().getName())
-                        || service.equals(exporter.getInvoker().getUrl().getPath())) {
-                    invoker = exporter.getInvoker();
-                    break;
-                }
-            }
-            if (invoker != null) {
-                Method[] methods = invoker.getInterface().getMethods();
-                for (Method method : methods) {
-                    if (buf.length() > 0) {
-                        buf.append("\r\n");
-                    }
-                    if (detail) {
-                        buf.append(ReflectUtils.getName(method));
-                    } else {
-                        buf.append(method.getName());
-                    }
-                }
-            } else {
-                buf.append("No such service " + service);
+            printSpecifiedService(service, buf, detail);
+
+            if (buf.length() == 0) {
+                buf.append("No such service: ").append(service);
             }
         }
         return buf.toString();
     }
 
+    private void printAllServices(StringBuilder buf, boolean detail) {
+        printAllProvidedServices(buf, detail);
+        printAllReferredServices(buf, detail);
+    }
+
+    private void printAllProvidedServices(StringBuilder buf, boolean detail) {
+        if (!ApplicationModel.allProviderModels().isEmpty()) {
+            buf.append("PROVIDER:\r\n");
+        }
+
+        for (ProviderModel provider : ApplicationModel.allProviderModels()) {
+            buf.append(provider.getServiceName());
+            if (detail) {
+                buf.append(" -> ");
+                buf.append(" published: ");
+                buf.append(isRegistered(provider.getServiceName()) ? "Y" : "N");
+            }
+            buf.append("\r\n");
+        }
+    }
+
+    private void printAllReferredServices(StringBuilder buf, boolean detail) {
+        if (!ApplicationModel.allConsumerModels().isEmpty()) {
+            buf.append("CONSUMER:\r\n");
+        }
+
+        for (ConsumerModel consumer : ApplicationModel.allConsumerModels()) {
+            buf.append(consumer.getServiceName());
+            if (detail) {
+                buf.append(" -> ");
+                buf.append(" addresses: ");
+                buf.append(getConsumerAddressNum(consumer.getServiceName()));
+            }
+        }
+    }
+
+    private void printSpecifiedService(String service, StringBuilder buf, boolean detail) {
+        printSpecifiedProvidedService(service, buf, detail);
+        printSpecifiedReferredService(service, buf, detail);
+    }
+
+    private void printSpecifiedProvidedService(String service, StringBuilder buf, boolean detail) {
+        for (ProviderModel provider : ApplicationModel.allProviderModels()) {
+            if (service.equalsIgnoreCase(provider.getServiceName())
+                    || service.equalsIgnoreCase(provider.getServiceInterfaceClass().getName())
+                    || service.equalsIgnoreCase(provider.getServiceInterfaceClass().getSimpleName())) {
+                buf.append(provider.getServiceName()).append(" (as provider):\r\n");
+                for (ProviderMethodModel method : provider.getAllMethods()) {
+                    printMethod(method.getMethod(), buf, detail);
+                }
+            }
+        }
+    }
+
+    private void printSpecifiedReferredService(String service, StringBuilder buf, boolean detail) {
+        for (ConsumerModel consumer : ApplicationModel.allConsumerModels()) {
+            if (service.equalsIgnoreCase(consumer.getServiceName())
+                    || service.equalsIgnoreCase(consumer.getServiceInterfaceClass().getName())
+                    || service.equalsIgnoreCase(consumer.getServiceInterfaceClass().getSimpleName())) {
+                buf.append(consumer.getServiceName()).append(" (as consumer):\r\n");
+                for (ConsumerMethodModel method : consumer.getAllMethods()) {
+                    printMethod(method.getMethod(), buf, detail);
+                }
+            }
+        }
+    }
+
+    private void printMethod(Method method, StringBuilder buf, boolean detail) {
+        if (detail) {
+            buf.append('\t').append(ReflectUtils.getName(method));
+        } else {
+            buf.append('\t').append(method.getName());
+        }
+        buf.append("\r\n");
+    }
 }
